fix(core): 修复中低优先级交互与查询问题
- 恢复审核后台商品管理查询并限制审核员访问
- 修正搜索通配符、分页边界和数据库时区配置
- 避免前端旧请求覆盖、分类误写及本地缓存异常
- 补齐冲突响应、回归测试并清理遗留备份源码

+++ v1/src/main/java/com/campustrade/platform/goods/mapper/GoodsMapper.java
@@ -34,7 +34,7 @@ public interface GoodsMapper {
                              @Param("categoryId") Long categoryId,
                              @Param("status") GoodsStatusEnum status,
                              @Param("limit") int limit,
-                             @Param("offset") int offset);
+                             @Param("offset") long offset);
 
     long countSearch(@Param("keywords") List<String> keywords,
                      @Param("categoryId") Long categoryId,
